+++ CHANGELOG.md
@@ -6,9 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## Added
+
+* `pyzet info` command which shows a bunch of stats about ZK repo
+
 ### Changed
 
 * Ask directly to continue if only one match was found when using grep patterns.
+
+### Removed
+
+* `pyzet tags --count`, as number of tags is now showed in `pyzet info`.
 
 ## [0.10.0] - 2024-01-17