+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.11.0] -- 2024-01-17
+
 ## Added
 
 * `pyzet info` command which shows a bunch of stats about ZK repo
@@ -156,7 +158,8 @@ Initial release.
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/tpwo/pyzet/releases/tag/v0.11.0
 [0.10.0]: https://github.com/tpwo/pyzet/releases/tag/v0.10.0
 [0.9.0]: https://github.com/tpwo/pyzet/releases/tag/v0.9.0
 [0.8.0]: https://github.com/tpwo/pyzet/releases/tag/v0.8.0